% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{detectMAFGenome}
\alias{detectMAFGenome}
\title{A function to detect MAF genome}
\usage{
detectMAFGenome(maf)
}
\arguments{
\item{maf}{The MAF object}
}
\value{
The list of object containing the genome
information
}
\description{
A function to detect MAF genome
}
\examples{
library(MAFDashRPackage)
#g<-detectMAFGenome(maf)
}
