% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMAF.R
\name{filterMAF}
\alias{filterMAF}
\title{Function to generate a dashboard from a MAF file.}
\usage{
filterMAF(
  maf_file,
  flag_genes = "default",
  save_name = NULL,
  no_filter = F,
  norm_alt_max = 1,
  t_alt_min = 1,
  t_depth_min = 20,
  tumor_freq_min = 0.05,
  norm_freq_max = 0.02,
  gnomAD_AF_max = 0.001,
  AF_max = 0.001,
  ExAC_AF_max = 0.001,
  n_callers = 2,
  variant_caller = NULL
)
}
\arguments{
\item{file}{The path of the file containing the mutation
information in the MAF format}
}
\value{
The output is the html file.
}
\description{
This function created an HTML file containing the
different figures and plots explaining the MAF dataset.
}
\examples{
library(MAFDashRPackage)
MAFfilePath <- system.file('extdata', 'test.maf', package = 'MAFDashRPackage')
#t <- getMAFDashboard(file = MAFfilePath)

}
\author{
Mayank Tondon, Ashish Jain
}
