% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCOSMICSignaturePlot.R
\name{generateCOSMICMutSigSimHeatmap}
\alias{generateCOSMICMutSigSimHeatmap}
\title{Computes and plot the cosine similarity of each individual signature is computed against each COSMIC signature
from COMIC V3.2}
\usage{
generateCOSMICMutSigSimHeatmap(
  mymaf,
  use_silent_mutations = FALSE,
  full_output = FALSE,
  show_broad_categories = TRUE,
  clin_data = NULL,
  clin_data_colors = NULL,
  add_sample_names = NULL,
  savename = NULL,
  fig_height = NULL,
  fig_width = NULL
)
}
\arguments{
\item{mymaf}{mutation count matrix (dimensions: a mutation features X n samples)}

\item{use_silent_mutations}{96 mutation count matrix (dimensions: a mutation features X m samples)}

\item{full_output}{return full output including the etiology matrix and plot data}

\item{show_broad_categories}{To show broad etiology categories}

\item{clin_data}{Clinical data to be plotted in the heatmap}

\item{clin_data_colors}{Clinical data colors}

\item{add_sample_names}{Whether or not to add column labels; if set to NULL, will add labels only if # samples less than 10}

\item{savename}{file name of the plot}

\item{fig_height}{Output height (inches); set to NULL to size automatically; only used if savename is set.}

\item{fig_width}{Output width (inches); set to NULL to size automatically; only used if savename is set.}
}
\value{
Complex Heatmap object. If full_output is TRUE it will consist of a list including heatmap object,
etiology matrix and plot data.
}
\description{
This function computes and plot the cosine similarity of each individual signature is computed against each COSMIC signature
from COMIC V3.2. The cosine similarity is a value between 0 (distinct) and 1 (identical) and indicates how much two vectors
are alike.
}
\examples{
library(MAFDash)
library(maftools)
library(ComplexHeatmap)
maf <- system.file("extdata", "test.mutect2.maf.gz", package = "MAFDash")
\donttest{val<-generateCOSMICMutSigSimHeatmap(read.maf(maf));draw(val)}


}
