% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOncoPlot.R
\name{generateOncoPlot}
\alias{generateOncoPlot}
\title{Function to generate a dashboard from a MAF file}
\usage{
generateOncoPlot(
  maf,
  cohort_freq_thresh = 0.01,
  auto_adjust_threshold = T,
  oncomat_only = F,
  clin_data = NULL,
  clin_data_colors = NULL
)
}
\arguments{
\item{maf}{The MAF object}

\item{cohort_freq_thresh}{Cohort Frequency Threshold}

\item{auto_adjust_threshold}{Auto Adjusted Threshold flag}

\item{oncomat_only}{Oncomat only flag}

\item{clin_data}{Clinical data}

\item{clin_data_colors}{Clinical data colors}
}
\value{
The onco plot object.
}
\description{
This function created an HTML file containing the
different figures and plots explaining the MAF dataset.
}
\examples{
library(MAFDashRPackage)
#g <- generateOncoPlot(maf)
#g

}
\author{
Mayank Tondon, Ashish Jain
}
