% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSingleRibbonPlot.R
\name{generateSingleRibbonPlot}
\alias{generateSingleRibbonPlot}
\title{Function to generate a dashboard from a MAF file.}
\usage{
generateSingleRibbonPlot(
  maf,
  onco_genes = NULL,
  save_name = NULL,
  ribbon_color = NULL,
  pval_high = 0.1,
  pval_low = 0.05,
  plot_frac_mut_axis = TRUE,
  rotate_plot_degrees = 0,
  shrink_factor = 1.3,
  scale_ribbon_to_fracmut = TRUE,
  sig_colors = NULL,
  gene_colors = NULL
)
}
\arguments{
\item{file}{The path of the file containing the mutation
information in the MAF format}
}
\value{
The output is the html file.
}
\description{
This function created an HTML file containing the
different figures and plots explaining the MAF dataset.
}
\examples{
library(MAFDashRPackage)
MAFfilePath <- system.file('extdata', 'test.maf', package = 'MAFDashRPackage')
#t <- getMAFDashboard(file = MAFfilePath)

}
\author{
Mayank Tondon, Ashish Jain
}
