% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTiTvPlot.R
\name{generateTiTvPlot}
\alias{generateTiTvPlot}
\title{Function to plot the frequency of Transitions and Transversions of gene mutations}
\usage{
generateTiTvPlot(
  maf,
  use_silent_mutations = FALSE,
  sampleOrder = NULL,
  color = NULL,
  showBarcodes = FALSE,
  textSize = 0.8,
  baseFontSize = 1,
  axisTextSize = c(1, 1),
  save_name = NULL
)
}
\arguments{
\item{maf}{A MAF object}

\item{use_silent_mutations}{Include synonymous variants in analysis. Defaults to FALSE.}

\item{sampleOrder}{Sample names in which the barplot should be ordered. Default NULL}

\item{color}{named vector of colors for each coversion class.}

\item{showBarcodes}{Whether to include sample names for barplot}

\item{textSize}{fontsize if showBarcodes is TRUE. Deafult 2.}

\item{baseFontSize}{font size. Deafult 1.}

\item{axisTextSize}{text size x and y tick labels. Default c(1,1).}

\item{save_name}{The name and path of the output file}
}
\value{
No return value. If 'save_name' is not provided, then the plot is printed to the current graphics device, otherwise a PDF is created at the given path.
}
\description{
This function plot the frequency of Transitions and Transversions of gene mutations
}
\examples{
library(MAFDash)
library(maftools)
maf <- system.file("extdata", "test.mutect2.maf.gz", package = "MAFDash")
plots<-generateTiTvPlot(read.maf(maf))

}
\author{
Mayank Tandon, Ashish Jain
}
