% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{generateVariantTable}
\alias{generateVariantTable}
\title{Make variant table from maf file}
\usage{
generateVariantTable(maf.filter, use_syn = F, extra_cols = c())
}
\arguments{
\item{maf.filter}{The MAF object}

\item{use_syn}{use_syn}

\item{extra_cols}{extra_cols}
}
\value{
Data table containing the variant information
}
\description{
Make variant table from maf file
}
\examples{
library(MAFDashRPackage)
#g<-generateVariantTable(maf)
}
