% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{geneSelectParser}
\alias{geneSelectParser}
\title{Function to the extact the Gene symbol from the input genes}
\usage{
geneSelectParser(genes_arg = NULL)
}
\arguments{
\item{genes_arg}{genes_arg}
}
\value{
A character vector containing the list of genes with gene symbols
}
\description{
Function to the extact the Gene symbol from the input genes
}
\examples{
library(MAFDash)
geneSelectParser()
}
