% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMAFDashboard.R
\name{getMAFDashboard}
\alias{getMAFDashboard}
\title{Function to generate a dashboard from a MAF file.}
\usage{
getMAFDashboard(
  filePath,
  outputFilePath,
  outputFileName = "output",
  outputFileTitle = "MAF Dash"
)
}
\arguments{
\item{filePath}{The path of the file containing the mutation
information in the MAF format}

\item{outputFilePath}{The path of the output html dashboard
file}

\item{outputFileName}{The name of the output html dashboard
file}

\item{outputFileTitle}{The title of the output html dashboard
file}
}
\value{
The dashboard html file
}
\description{
This function created an HTML file containing the
different figures and plots explaining the MAF dataset.
}
\examples{
library(MAFDashRPackage)
#MAFfilePath <- system.file('extdata', 'test.maf', package = 'MAFDashRPackage')
#t <- getMAFDashboard(file = MAFfilePath)

}
\author{
Mayank Tondon, Ashish Jain
}
