% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMAFDataTCGA.R
\name{getTCGAClinicalColors}
\alias{getTCGAClinicalColors}
\title{Makes reasonable colors for some TCGA clinical annoations}
\usage{
getTCGAClinicalColors(ageRange = c(0, 100))
}
\arguments{
\item{ageRange}{The range of patient's age to generate color vector from RColorBrewer}
}
\value{
A list containing the TCGA clinical annotations
}
\description{
This will return a list of colors that can be used with TCGA clinical annotations
}
\examples{
library(MAFDash)
colorList <- getTCGAClinicalColors()
}
\author{
Mayank Tandon, Ashish Jain
}
