% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{make_column_annotation}
\alias{make_column_annotation}
\title{Make the annotation data frame from the TCGA clinical dataset}
\usage{
make_column_annotation(my_clin_dat, names_to_match, my_colors = NULL)
}
\arguments{
\item{my_clin_dat}{Clinical dataset in a data frame}

\item{names_to_match}{The list containing the matched patient's name}

\item{my_colors}{my_colors}
}
\value{
The clinical annotation data
}
\description{
This function creates a annotation data frame from the TCGA clinical dataset
}
\examples{
library(MAFDash)
}
