% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{make_overlap_plot}
\alias{make_overlap_plot}
\title{Function to generate a dashboard from a MAF file.}
\usage{
make_overlap_plot(
  mymaf,
  use_silent_mutations = F,
  summarize_by = "gene",
  plotType = c("ribbon", "heatmap"),
  savename = "overlap_plot.pdf",
  savewidth = 8,
  saveheight = 8
)
}
\arguments{
\item{file}{The path of the file containing the mutation
information in the MAF format}
}
\value{
The output is the html file.
}
\description{
This function created an HTML file containing the
different figures and plots explaining the MAF dataset.
}
\examples{
library(MAFDashRPackage)
MAFfilePath <- system.file('extdata', 'test.maf', package = 'MAFDashRPackage')
#t <- getMAFDashboard(file = MAFfilePath)

}
\author{
Mayank Tondon, Ashish Jain
}
