% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{make_variant_table}
\alias{make_variant_table}
\title{Make variant table from maf file}
\usage{
make_variant_table(maf.filter, use_syn = F, extra_cols = c())
}
\description{
Make variant table from maf file
}
\keyword{internal}
