% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{my_mutation_colors}
\alias{my_mutation_colors}
\title{Returns the colors for each mutation}
\usage{
my_mutation_colors()
}
\value{
The mutation color data frame
}
\description{
This function returns the colors for each mutation
}
\examples{
library(MAFDashRPackage)
}
