% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{oncoplot_annotation_func}
\alias{oncoplot_annotation_func}
\title{Returns the mutation colors for oncoplot function}
\usage{
oncoplot_annotation_func()
}
\value{
The mutation color data frame
}
\description{
This function returns the mutation colors for oncoplot function
}
\examples{
library(MAFDashRPackage)
oncoplot_annotation_func()
}
