% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{createOncoMatrix}
\alias{createOncoMatrix}
\title{Creates matrix for oncoplot from maf file}
\usage{
createOncoMatrix(m, g = NULL, chatty = TRUE, add_missing = FALSE)
}
\description{
Cretaes matrix for oncoplot from maf file
}
