% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlots.R
\name{generateOncoplot}
\alias{generateOncoplot}
\title{Function to generate a dashboard from a MAF file.}
\usage{
generateOncoplot(
  maf.filtered,
  cohort_freq_thresh = 0.01,
  auto_adjust_threshold = T,
  oncomat_only = F,
  clin_data = NULL,
  clin_data_colors = NULL
)
}
\arguments{
\item{file}{The path of the file containing the mutation
information in the MAF format}
}
\value{
The output is the html file.
}
\description{
This function created an HTML file containing the
different figures and plots explaining the MAF dataset.
}
\examples{
library(MAFDashRPackage)
MAFfilePath <- system.file('extdata', 'test.maf', package = 'MAFDashRPackage')
#t <- getMAFDashboard(file = MAFfilePath)

}
\author{
Mayank Tondon, Ashish Jain
}
