% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMAFDataTCGA.R
\name{getMAFdataTCGA}
\alias{getMAFdataTCGA}
\title{Function to generate a dashboard from a MAF file.}
\usage{
getMAFdataTCGA(
  tcga_dataset = "ACC",
  save_folder = file.path("data"),
  variant_caller = "mutect2"
)
}
\arguments{
\item{file}{The path of the file containing the mutation
information in the MAF format}
}
\value{
The output is the html file.
}
\description{
This function created an HTML file containing the
different figures and plots explaining the MAF dataset.
}
\examples{
library(MAFDashRPackage)
MAFfilePath <- system.file('extdata', 'test.maf', package = 'MAFDashRPackage')
#t <- getMAFDashboard(file = MAFfilePath)

}
\author{
Mayank Tondon, Ashish Jain
}
