% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMAFDataTCGA.R
\name{getMAFdataTCGA}
\alias{getMAFdataTCGA}
\title{Function to extract the mutation data in MAF format from TCGA}
\usage{
getMAFdataTCGA(
  cancerCode = "ACC",
  outputFolder = file.path("data"),
  variant_caller = "mutect2"
)
}
\arguments{
\item{cancerCode}{The TCGA cancer code}

\item{outputFolder}{The path of the file containing the mutation
information in the MAF format}
}
\value{
A list containing path of mutation annotation file
}
\description{
This function download and extract the mutation
 data in MAF format from TCGA.
}
\examples{
library(MAFDashRPackage)
cancerCode <- "ACC"
outputFolderPath <- "."
#maf <- getMAFdataTCGA(cancerCode = cancerCode,outputFolder = outputFolderPath)
}
\author{
Mayank Tondon, Ashish Jain
}
