% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMAFDataTCGA.R
\name{getMAFdataTCGA}
\alias{getMAFdataTCGA}
\title{Function to extract the mutation data in MAF format from TCGA}
\usage{
getMAFdataTCGA(
  cancerCodes = c("ACC"),
  outputFolder = tempdir(),
  variant_caller = "mutect2"
)
}
\arguments{
\item{cancerCodes}{A character vector of TCGA cancer codes}

\item{outputFolder}{path of the file containing the mutation
information in the MAF format}

\item{variant_caller}{type of variant caller in TCGA}
}
\value{
A vector containing the path of the downloaded mutation annotation files
}
\description{
This function download and extract the mutation
 data in MAF format from TCGA.
}
\examples{
library(MAFDash)
cancerCodes <- c("ACC","THCA")
outputFolderPath <- tempdir()
#maf <- getMAFdataTCGA(cancerCodes = cancerCodes,outputFolder = outputFolderPath)
}
\author{
Mayank Tandon, Ashish Jain
}
