% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMAFDataTCGA.R
\name{getTCGAClinicalAnnotation}
\alias{getTCGAClinicalAnnotation}
\title{Function to extract the clinical annotations from TCGA}
\usage{
getTCGAClinicalAnnotation(
  cancerCode = "ACC",
  outputFolder = file.path("data"),
  plotdata = NULL
)
}
\arguments{
\item{cancerCode}{The TCGA cancer code}

\item{outputFolder}{The path of the file containing the clinical
annotations from TCGA}

\item{plotdata}{Flag to plot the annotations}
}
\value{
A list containing the TCGA clinical annotations
}
\description{
This function download and extract the clinical
annotations from TCGA.
}
\examples{
library(MAFDash)
cancerCode <- "ACC"
outputFolderPath <- "."
#maf <- getMAFdataTCGA(cancerCode = cancerCode,outputFolder = outputFolderPath)

}
\author{
Mayank Tandon, Ashish Jain
}
